+++ examples/resources/bluecat_host_record/resource.tf
@@ -1,5 +1,5 @@
 resource "bluecat_host_record" "hostname" {
-  view      = data.bluecat_entity.view.id
+  view_id   = data.bluecat_entity.view.id
   name      = "hostname"
   dns_zone  = "example.com"
   addresses = ["192.168.1.100"]