+++ internal/provider/resource_ip4_address.go
@@ -50,6 +50,7 @@ func resourceIP4Address() *schema.Resource {
 				Description: "A map of all custom properties associated with the IPv4 address.",
 				Type:        schema.TypeMap,
 				Optional:    true,
+				Computed:    true,
 				Elem: &schema.Schema{
 					Type: schema.TypeString,
 				},
@@ -87,6 +88,7 @@ func resourceIP4Address() *schema.Resource {
 func resourceIP4AddressCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	mutex.Lock()
 	client := meta.(*apiClient).Client
+	client.Login(meta.(*apiClient).Username, meta.(*apiClient).Password)
 
 	configID, err := strconv.ParseInt(d.Get("configuration_id").(string), 10, 64)
 	if err = gobam.LogoutClientIfError(client, err, "Unable to convert configuration_id from string to int64"); err != nil {
@@ -136,6 +138,7 @@ func resourceIP4AddressCreate(ctx context.Context, d *schema.ResourceData, meta 
 func resourceIP4AddressRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	mutex.Lock()
 	client := meta.(*apiClient).Client
+	client.Login(meta.(*apiClient).Username, meta.(*apiClient).Password)
 
 	id, err := strconv.ParseInt(d.Id(), 10, 64)
 	if err = gobam.LogoutClientIfError(client, err, "Unable to convert id from string to int64"); err != nil {
@@ -185,6 +188,7 @@ func resourceIP4AddressRead(ctx context.Context, d *schema.ResourceData, meta in
 func resourceIP4AddressUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	mutex.Lock()
 	client := meta.(*apiClient).Client
+	client.Login(meta.(*apiClient).Username, meta.(*apiClient).Password)
 
 	id, err := strconv.ParseInt(d.Id(), 10, 64)
 	if err = gobam.LogoutClientIfError(client, err, "Unable to convert id from string to int64"); err != nil {
@@ -234,6 +238,7 @@ func resourceIP4AddressUpdate(ctx context.Context, d *schema.ResourceData, meta 
 func resourceIP4AddressDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	mutex.Lock()
 	client := meta.(*apiClient).Client
+	client.Login(meta.(*apiClient).Username, meta.(*apiClient).Password)
 
 	id, err := strconv.ParseInt(d.Id(), 10, 64)
 	if err = gobam.LogoutClientIfError(client, err, "Unable to convert id from string to int64"); err != nil {